requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
requests>=2.31.0
pdfplumber>=0.10.3
python-docx>=1.1.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
numpy>=1.26.0
